Create red_blu_stack.py
Design an ADT for a two-color, double-stack ADT that consists of two stacks—one “red” and one “blue”—and has as its operations color-coded versions of the regular stack ADT operations. For example, this ADT should support both a red push operation and a blue push operation. Give an efficient implementation of this ADT using a single array whose ca- pacity is set at some value N that is assumed to always be larger than the sizes of the red and blue stacks combined.